Fixing missing msg in the flow.
After the message is processed successfully, it is explictly delete from the queue/batch. If the message is not processed, it will remain in the batch, handle the error, and continue processing the remaining messages from the batch. At the end of the batch processing, if there are any unprocessed message due to error, throw exception at the end, so that they will go to the DLQ.